Create ChuaEncryption.py
This takes in a string that needs to be encrypted. This algorithm encrypts the string by using the noise and floating points of a electrical circuit named after Dr. Leon Chua. The results have position data points. This is mainly for written work. You can look at EncyptionWave.py to see the string encrypted into a wave form for signal submission.